Add support for extracting `body-calendar`
Mailgun extracts calendar invites that are embedded in emails sent
from Outlook and adds them in to a param called `body-calendar`,
but griddler does not permit this in the params.

This will extract the information in the `body-calendar` param,
transform it in to an `ActionDispatch::Http::UploadedFile`, & then
add it to the `:attachments` param.